URI encode/decode query map in Hoplon UI routes
If we have a string in the query map and use such a URL
in the href attribute of an anchor within an HTML email,
then the double quotes should be turned into HTML entities,
so within the source of the email body it looks like:

<a href="http://tomplus.local:8100/#/verify-access?code=&quot;eyJhbG...riV4&quot;">link</a>

Such HTML entities are decoded differently by different mail clients though.

Mailhog does not decode them:

<a href="http://tomplus.local:8100/#/verify-access?code=&quot;eyJhbG...u6o&quot;"
     target="_blank">
        link
</a>

But Gmail turns it into this monster:

<a href="https://single.page.app/#/verify-access?code=%22eyJhbG...bsb0%22"
     target="_blank"
     data-saferedirecturl="https://www.google.com/url?q=https://single.page.app/%23/verify-access?code%3D%2522eyJhbG...bsb0%2522&amp;source=gmail&amp;ust=1474623455271000&amp;usg=AFQjCNEyjUDQOFLtqVBxmqq8u1ZqVGzVWg">
         link
</a>

Note how the "&" signs within the saferedirecturl are HTML entities but the
value of the "q" URL query param is double URI encoded.